Handle hashcat status line less rigidly
In Hashcat 7.0.0, there is an extra field POWER which tells you how much
power each device is drawing. This currently breaks the Hashtopia
agent-python when using 7.0.0 because it expects the end of the line to
be integer values for UTIL. In order to avoid this problem in the future,
be less rigid in parsing how the status line is returned. Print out
messages for any unexpected fields.